sched/topology: Add update_domain_cpu()
Introduce update_domain_cpu(), which is a wrapper of
update_top_cache_domain(). In update_domain_cpu() the
cpu hotplug lock is to protect against the rebuild of
sched domain, and the rcu read lock is to protect against
the dereference of domain tree(rq->sd) in update_top_cache_domain().
This patch is to prepare for the next patch to update the
flags of sched domain via sysctl.

No intentional functional impact.

link: https://lore.kernel.org/all/d06455426aecadd7c7751ed0a5bd24d1bd54422c.1594062828.git.yu.c.chen@intel.com/
Signed-off-by: Chen Yu <yu.c.chen@intel.com>
Signed-off-by: guojinhui.liam <guojinhui.liam@bytedance.com>