x86/fault: Correct a few user vs kernel checks wrt WRUSS
commit 56e62cd28aaae2fcbec8af67b05843c47c6da170 upstream.

In general, page fault errors for WRUSS should be just like get_user(),
etc.  Fix three bugs in this area:

There is a comment that says that, if the kernel can't handle a page fault
on a user address due to OOM, the OOM-kill-and-retry logic would be
skipped.  The code checked kernel *privilege*, not kernel mode, so it
missed WRUSS.  This means that the kernel would malfunction if it got OOM
on a WRUSS fault -- this would be a kernel-mode, user-privilege fault, and
the OOM killer would be invoked and the handler would retry the faulting
instruction.

A failed user access from kernel while a fatal signal is pending should
fail even if the instruction in question was WRUSS.

do_sigbus() should not send SIGBUS for WRUSS -- it should handle it like
any other kernel mode failure.

Intel-SIG: commit 56e62cd28aaa x86/fault: Correct a few user vs kernel
checks wrt WRUSS.
Backport summary: This is preparation patch for CET.

Signed-off-by: Andy Lutomirski <luto@kernel.org>
Signed-off-by: Borislav Petkov <bp@suse.de>
Link: https://lkml.kernel.org/r/a7b7bcea730bd4069e6b7e629236bb2cf526c2fb.1612924255.git.luto@kernel.org
[ Jing Han amend commit log ]
Signed-off-by: Jing Han <jing1.han@intel.com>
Signed-off-by: Gang Li <ligang.bdlg@bytedance.com>