x86/fault: Improve kernel-executing-user-memory handling
commit 03c81ea3331658f613bb2913d33764a4e0410cbd upstream.

Right now, the case of the kernel trying to execute from user memory
is treated more or less just like the kernel getting a page fault on a
user access. In the failure path, it checks for erratum #93, tries to
otherwise fix up the error, and then oopses.

If it manages to jump to the user address space, with or without SMEP,
it should not try to resolve the page fault. This is an error, pure and
simple. Rearrange the code so that this case is caught early, check for
erratum #93, and bail out.

 [ bp: Massage commit message. ]

Intel-SIG: commit 03c81ea33316 x86/fault: Improve
kernel-executing-user-memory handling.
Backport summary: This is preparation patch for CET.

Signed-off-by: Andy Lutomirski <luto@kernel.org>
Signed-off-by: Borislav Petkov <bp@suse.de>
Link: https://lkml.kernel.org/r/ab8719c7afb8bd501c4eee0e36493150fbbe5f6a.1612924255.git.luto@kernel.org
[ Jing Han amend commit log ]
Signed-off-by: Jing Han <jing1.han@intel.com>
Signed-off-by: Gang Li <ligang.bdlg@bytedance.com>