KVM: x86/mmu: Ensure MMU pages are available when allocating roots
commit 6e6ec58485746eb64487bd49bf5cd90ded3d2cf6 upstream.

backport note:
This commit should be eariler than commit
4a38162ee9f10f5f67c36f4f5aa4f6be2657efd5, and the latter depends on it.
But the latter has been ported with kvm_mmu_load() part changes missed.
So apply the missed changes with this chance to sync to upstream.

Hold the mmu_lock for write for the entire duration of allocating and
initializing an MMU's roots.  This ensures there are MMU pages available
and thus prevents root allocations from failing.  That in turn fixes a
bug where KVM would fail to free valid PAE roots if a one of the later
roots failed to allocate.

Add a comment to make_mmu_pages_available() to call out that the limit
is a soft limit, e.g. KVM will temporarily exceed the threshold if a
page fault allocates multiple shadow pages and there was only one page
"available".

Note, KVM _still_ leaks the PAE roots if the guest PDPTR checks fail.
This will be addressed in a future commit.

Cc: Ben Gardon <bgardon@google.com>
Signed-off-by: Sean Christopherson <seanjc@google.com>
Message-Id: <20210305011101.3597423-7-seanjc@google.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: Jiawei Gu <gujiawei.will@bytedance.com>