x86/fpu: Use fpstate in __copy_xstate_to_uabi_buf()
commit 3ac8d75778fc8c1c22daad9bc674166b862f6f6e upstream.

With dynamically enabled features the copy function must know the features
and the size which is valid for the task. Retrieve them from fpstate.

Intel-SIG: commit 3ac8d75778fc x86/fpu: Use fpstate in
__copy_xstate_to_uabi_buf().
The second part of the effort to support AMX: Move register state into
a container struct.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Borislav Petkov <bp@suse.de>
Link: https://lkml.kernel.org/r/20211013145323.181495492@linutronix.de
[ Lin Wang: amend commit log ]
Signed-off-by: Lin Wang <lin.x.wang@intel.com>
Signed-off-by: Gang Li <ligang.bdlg@bytedance.com>