x86/fpu: Move xstate size to fpu_*_cfg
commit 2bd264bce238cedbf00bde1f28ad51ba45b9114e upstream.

Use the new kernel and user space config storage to store and retrieve the
XSTATE buffer sizes. The default and the maximum size are the same for now,
but will change when support for dynamically enabled features is added.

Intel-SIG: commit 2bd264bce238 x86/fpu: Move xstate size to fpu_*_cfg.
The third part of the effort to support AMX: consolidate the size and
feature information.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Borislav Petkov <bp@suse.de>
Link: https://lkml.kernel.org/r/20211014230739.296830097@linutronix.de
[ Lin Wang: amend commit log ]
Signed-off-by: Lin Wang <lin.x.wang@intel.com>
Signed-off-by: Gang Li <ligang.bdlg@bytedance.com>