bytedance: bpf: Optimized compatibility with private helper functions
Since the private helper function (unsafe_helper) doesn't merge to the
kernel of the community. In order to ensure that the order of the new
helper functions which backported from community kernel is consistent,
Now the private helper function needs to be placed last. But it's not
good enough.
So set BPF_FUNC_unsafe_helper between INT_MAX/2 and INT_MAX away from
the __BPF_FUNC_MAX_ID.

Signed-off-by: Feng Zhou <zhoufeng.zf@bytedance.com>