downstream: mbigen: vtimer mbigen driver support
This device is designed as the accessory device of mbigen to assist vtimer
to support irqbypass capability. It can translate the level vtimer irq
which percpu vtimer triggered to MSG interrupt (routed by ITS, and inject
to cpu).

Multiple mbigens may exist on one CPU Die and not all mbigens are related
to vtimer. Use MBIX_VPPI_ITS_TA to determin target peri mbigen related to
the vtimer. For each cpu, we calculate the offset of the corresponding GICR
phys addr to get the correct MBIGEN pin.

This implemented:
(1) vtimer mbigen device probed and inited, which indicate the device as
    the vtimer accessory dev.

(2) open some device configuration interfaces for hyper to configure.

Signed-off-by: Nianyao Tang <tangnianyao@huawei.com>
Signed-off-by: Zenghui Yu <yuzenghui@huawei.com>
Signed-off-by: Kunkun Jiang <jiangkunkun@huawei.com>
Signed-off-by: Dongxu Sun <sundongxu3@huawei.com>
Signed-off-by: Slim6882 <15605196882@163.com>
Signed-off-by: Qi Xi <xiqi2@huawei.com>