x86/resctrl: Add a helper to avoid reaching into the arch code resource list
commit 3c021531131c9ee5df5da739819a515326ee9570 upstream.

Resctrl occasionally wants to know something about a specific resource.
It reaches into the arch code's rdt_resources_all.

Once the filesystem parts of resctrl are moved to /fs/, this means it
will need visibility of the architecture specific structure.

Move the level enum to the resctrl header and add a helper to retrieve
the resctrl struct by id. resctrl_arch_get_resource() should not return
NULL for any value in the enum, it may instead return a dummy resource
that is !alloc_enabled && !mon_enabled.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Zeng Heng <zengheng4@huawei.com>