downstream: drivers/perf: hisi: Add support for L3C PMU v3
This patch adds support for L3C PMU v3. The v3 L3C PMU supports
an extended events space which can be controlled in a second
address space with a separate overflow interrupt. The offset
of the control/event registers keep the same. The extended events
with original ones together cover the monitoring of all the L3C
transactions.

The extended events is specified with `ext[=1]` option for the
driver to distinguish like:
perf stat -e hisi_sccl0_l3c0_0/event=<event_id>,ext/

Currently only event option using config bit [7, 0]. There's
still lots of field unused. Make ext using config 16 and
reserve bit [15, 8] for event option for future extension.

The hw_perf_event::event_base is initialized to the base MMIO
address of the event and will be used for later control,
overflow handling and counts readout.

We still make use of the Uncore PMU framework for handling the
events and interrupt migration on CPU hotplug. The framework's
cpuhp callback will handle the event migration and interrupt
migration of orginial event, if PMU supports extended events
then the interrupt of extended events is migrated to the same
CPU choosed by the framework.

A new HID of HISI0215 is used for this version of L3C PMU.

Signed-off-by: Yicong Yang <yangyicong@hisilicon.com>
Signed-off-by: Slim6882 <15605196882@163.com>
Signed-off-by: yeyiyang <yeyiyang3@h-partners.com>