downstream: kvm: arm64: avoid sending multi-SGIs in IPIV
IPIV optimizes unicast scenarios and does not support multicast or
broadcast. To prevent GuestOS from generating multicast and broadcast,
the MPIDR  of vCPUs is modified to ensure that the [aff3, aff2, aff1]
fields are uniquely assigned for each vCPU within a virtual machine,
while all aff0 fields are set to zero. This configuration guarantees
the uniqueness of vCPU affinity identifiers at the architecture level,
thereby suppressing the generation of SGI multicast and broadcast
signals by the GuestOS.

Signed-off-by: Jinqian Yang <yangjinqian1@huawei.com>
Signed-off-by: Slim6882 <15605196882@163.com>
Signed-off-by: chenyi <chenyi211@h-partners.com>