downstream: fs/resctrl: Fix kmemleak caused by closid_init()
When the resctrl fs is mounted, closid_init() will allocate memory of
closid_free_map, but when the fs is unmounted, it is never released,
which will cause a memory leak.

Add release of closid_free_map in rdt_kill_sb() to ensure no leaks when
the fs is unmounted.

Fixes: 1e7284b7981c ("fs/resctrl: Remove the limit on the number of CLOSID")
Signed-off-by: Zeng Heng <zengheng4@huawei.com>