downstream: arm64/mpam: Add mbw_min and cmin features
Implement mbw_min/cmin feature for mata/cache msc, and provide MBMIN/LxMIN
interface in schemata file for user.

On some specific platforms, L3 implements the cmin feature, but L2 may not
provide it, or vice versa. Therefore, it is necessary to separately
determine whether the cache msc has implemented the corresponding feature.

Before transmitting the cache MIN portion value to msc, need to check
whether the CMIN setting from user space is validate or not. The same
goes for MBMIN.

It should be noted that the new configuration value is in decimal form
instead of hexadecimal, so here reuse the parse_bw() to analysis input
value for cache msc.

Signed-off-by: Zeng Heng <zengheng4@huawei.com>