downstream: arm64/mpam: Add PRIO feature
The Priority Partition feature controls the priority of requests attributed
to the target PARTID.

In addition to the MATA msc implementing priority features, the L3
controller can also have the same features. Here the *MBPRI* and *L3PRI*
interface are provided in the schemata file to allow users to configure
and query QoS partition settings.

Before transmitting the priority value to msc, it needs to check whether
the setting from user is validate or not.

Signed-off-by: Zeng Heng <zengheng4@huawei.com>