downstream: KVM: arm64: using kvm_vgic_global_state for ipiv
when kvm-arm.vgic_v4_enable=0 kvm-arm.ipiv_enabled=1 is configured in the
cmdline, the Host KVM displays "ipiv enabled", but ipiv is not enabled.
gic_data.rdists.has_rvpeid is hardware-level information, which does not
indicate that GICv4.1 is enabled for KVM.

So whether the host supports IPIV depends on KVM enables GICv4.1m instead
of hardware supports GICv4.1.

Signed-off-by: Jinqian Yang <yangjinqian1@huawei.com>
Signed-off-by: nzq-hw <nizhiqiang1@huawei.com>